Move uuid utils to separate file - it wasn't possible to import the utils file in e2e tests without it
